requirements.txt
# FastAPI Core
fastapi>=0.100.0,<0.105.0
uvicorn[standard]>=0.23.0,<0.25.0
python-multipart>=0.0.6

# AI/ML Core
openai>=1.3.0,<2.0.0
pillow>=9.0.0,<11.0.0

# Data Processing
pandas>=1.5.0,<2.1.0
openpyxl>=3.1.0
pydantic>=1.10.0,<2.0.0

# Background Tasks & Async
aiofiles>=22.0.0,<24.0.0

# Configuration & Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
pathlib2>=2.3.0
typing-extensions>=4.0.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
pytest-mock>=3.12.0
coverage>=7.3.0
black>=22.0.0
flake8>=4.0.0,<6.0.0
mypy>=0.991
